requirements.txt
transformers
sentencepiece
accelerate
datasets
rouge_score
torch 
fastapi
uvicorn
python-multipart
python-docx